+++ GamePortal/GamePortal/Models/ApplicationDbContext.cs
@@ -14,6 +14,7 @@ namespace GamePortal.Models
         public DbSet<Category> Categories { get; set; }
         public DbSet<News> News { get; set; }
         public DbSet<Purchase> Purchases { get; set; }
+        public DbSet<CartItem> CartItems { get; set; }
 
         //Для указания точности для типа данных Decimal
         protected override void OnModelCreating(ModelBuilder modelBuilder)
@@ -23,6 +24,10 @@ namespace GamePortal.Models
             modelBuilder.Entity<Game>()
                 .Property(g => g.Price)
                 .HasPrecision(18, 2);
+
+            modelBuilder.Entity<Game>()
+                .Property(g => g.IsActive)
+                .HasDefaultValue(true);
         }
     }